refactor: rename fn
