+++ packages/v4-sdk/src/functions/bin/getAmounts.ts
@@ -2,7 +2,7 @@ import { SCALE_OFFSET } from '../../constants/binPool'
 import { calculateSwapFee } from '../../utils/calculateSwapFee'
 import { mulShiftRoundDown, mulShiftRoundUp } from '../../utils/math/mulShift'
 import { shiftDivRoundDown, shiftDivRoundUp } from '../../utils/math/shiftDiv'
-import { BinPoolState } from './getBinPool'
+import { BinPoolState } from './createBinPool'
 import { getFeeAmount } from './getFeeAmount'
 import { getFeeAmountFrom } from './getFeeAmountFrom'
 import { getPriceFromId } from './getPriceFromId'
