+++ packages/v4-sdk/src/functions/bin/index.ts
@@ -1,6 +1,6 @@
+export * from './createBinPool'
 export * from './getAmounts'
 export * from './getBase'
-export * from './getBinPool'
 export * from './getExternalFeeAmt'
 export * from './getFeeAmount'
 export * from './getFeeAmountFrom'
