+++ packages/v4-sdk/src/functions/bin/swap.ts
@@ -6,8 +6,8 @@ import { SCALE_OFFSET } from '../../constants/binPool'
 import { calculateSwapFee } from '../../utils/calculateSwapFee'
 import { mulShiftRoundUp } from '../../utils/math/mulShift'
 import { shiftDivRoundUp } from '../../utils/math/shiftDiv'
+import { BinPoolState } from './createBinPool'
 import { getAmounts } from './getAmounts'
-import { BinPoolState } from './getBinPool'
 import { getExternalFeeAmt } from './getExternalFeeAmt'
 import { getFeeAmount } from './getFeeAmount'
 import { getNextNonEmptyBin } from './getNextNonEmptyBin'
